# apps/backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers": {
      "get": {
        "operationId": "CustomerController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CustomerProfileResponseDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/MetaDataDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Find all Customers",
        "tags": [
          "Customer"
        ]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "CustomerController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CustomerProfileResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get a specific customer by ID",
        "tags": [
          "Customer"
        ]
      },
      "patch": {
        "operationId": "CustomerController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "nullable": true
                  },
                  "fullName": {
                    "type": "string",
                    "nullable": true
                  },
                  "phone": {
                    "type": "string",
                    "nullable": true
                  },
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  },
                  "customer[loyaltyPoints]": {
                    "type": "number",
                    "nullable": true
                  },
                  "customer[defaultAddress][street]": {
                    "type": "string",
                    "nullable": true
                  },
                  "customer[defaultAddress][city]": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CustomerProfileResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a customer",
        "tags": [
          "Customer"
        ]
      },
      "delete": {
        "operationId": "CustomerController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CustomerProfileResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a customer",
        "tags": [
          "Customer"
        ]
      }
    },
...
    "/dashboard/overview": {
      "get": {
        "operationId": "DashboardController_getOverview",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DashboardOverviewDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get KPI overview",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/orders-by-status": {
      "get": {
        "operationId": "DashboardController_getOrdersByStatus",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {}
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Order count grouped by status",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/revenue-by-day": {
      "get": {
        "operationId": "DashboardController_getRevenueByDay",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {}
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Daily revenue for date range",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/top-items": {
      "get": {
        "operationId": "DashboardController_getTopItems",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {}
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Most ordered menu items",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/staff-performance": {
      "get": {
        "operationId": "DashboardController_getStaffPerformance",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {}
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Orders served per waiter",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/table-utilization": {
      "get": {
        "operationId": "DashboardController_getTableUtilization",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GenericResponseDto"
                    },
                    {
                      "properties": {}
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Table occupancy rates",
        "tags": [
          "Dashboard"
        ]
      }
    }
  },
  "info": {
    "title": "milky",
    "description": "The milky API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "MetaDataDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 10
          },
          "total": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
      "GenericResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDataDto"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "AddressDto": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "example": "123 Main St"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "postalCode": {
            "type": "string",
            "example": "10001"
          },
          "country": {
            "type": "string",
            "example": "USA"
          }
        },
        "required": [
          "street",
          "city",
          "country"
        ]
      },
      "CustomerDto": {
        "type": "object",
        "properties": {
          "loyaltyPoints": {
            "type": "number",
            "example": 0
          },
          "defaultAddress": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "preferences": {
            "type": "object",
            "example": {
              "dietaryRequirements": "vegan"
            }
          }
        }
      },
...
      "CustomerProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "60d5ecb74d6bb8996032cb0a"
          },
          "email": {
            "type": "string",
            "example": "customer@example.com"
          },
          "fullName": {
            "type": "string",
            "example": "Jane Customer"
          },
          "phone": {
            "type": "string",
            "example": "+1234567890"
          },
          "globalRole": {
            "type": "string",
            "example": "CUSTOMER"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "globalRole"
        ]
      },
...
      "DashboardOverviewDto": {
        "type": "object",
        "properties": {
          "totalOrders": {
            "type": "number"
          },
          "totalRevenue": {
            "type": "number"
          },
          "avgOrderValue": {
            "type": "number"
          },
          "activeTables": {
            "type": "number"
          }
        },
        "required": [
          "totalOrders",
          "totalRevenue",
          "avgOrderValue",
          "activeTables"
        ]
      }
    }
  }
}